자신의 게시물일 경우에만 DRAFT 보기, 수정, 삭제 가능

+++ app/blueprints/page/views.py
@@ -116,6 +116,12 @@ def post_detail_slug(slug):
         .filter(Post.slug == slug) \
         .first_or_404()
 
+    if post.status != Post.STATUS_PUBLIC \
+        and not current_user.is_authenticated \
+            or current_user.id != post.author_id:
+        return render_template(current_app.config['RABIANG_SITE_THEME'] +
+                               '/404.html'), 404
+
     form = CommentForm()
 
     if form.validate_on_submit():
@@ -167,6 +173,12 @@ def post_detail_slug(slug):
 @permission_required('post', 'view')
 def post_detail_id(post_id):
     post = Post.query.get_or_404(post_id)
+
+    if post.status != Post.STATUS_PUBLIC \
+            and not current_user.is_authenticated \
+            or current_user.id != post.author_id:
+        return render_template(current_app.config['RABIANG_SITE_THEME'] +
+                               '/404.html'), 404
 
     form = CommentForm()
 
@@ -267,6 +279,10 @@ def post_create():
 def post_edit(post_id):
     post = Post.query.get_or_404(post_id)
 
+    if not current_user.is_authenticated or current_user.id != post.author_id:
+        return render_template(current_app.config['RABIANG_SITE_THEME'] +
+                               '/404.html'), 404
+
     categories = build_tree_tuple_list(PageCategory, prefix=True)
 
     form = PostForm(obj=post)
@@ -320,6 +336,10 @@ def post_edit(post_id):
 @permission_required('post', 'delete')
 def post_delete(post_id):
     post = Post.query.get_or_404(post_id)
+
+    if not current_user.is_authenticated or current_user.id != post.author_id:
+        return render_template(current_app.config['RABIANG_SITE_THEME'] +
+                               '/404.html'), 404
 
     form = DeletePostForm()
